backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.1
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1